Unify CustomAnalogClock and CustomAnalogClock4
