+++ res/layout-v14/analog_clock_layout_4.xml
@@ -8,7 +8,7 @@
             time
 
     -->
-    <com.firebirdberlin.nightdream.CustomAnalogClock4 xmlns:android="http://schemas.android.com/apk/res/android"
+    <com.firebirdberlin.nightdream.CustomAnalogClock xmlns:android="http://schemas.android.com/apk/res/android"
         android:id="@+id/analog_clock"
         android:layout_width="200dp"
         android:layout_height="200dp"
